requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altgraph==0.17.4
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
packaging==23.2
pefile==2023.2.7
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
pytest~=8.0.2
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
requests==2.31.0
setuptools==69.1.0
urllib3==2.2.0
yarl==1.9.4

